pyproject.toml
[project]
name = "travel-log"
version = "0.1.0"
description = "travel_log"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "svlearn-bootcamp>=0.1.7",
    "deepface>=0.0.93",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "numpy>=1.24.0",
    "tf-keras>=2.20.0",
    "streamlit>=1.28.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "requests>=2.31.0",
    "dspy-ai>=3.0.4",
    "qdrant-client>=1.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    # ... other includes like README, pyproject.toml, etc.
]

[tool.hatch.build.targets.wheel]
packages = ["src/travel_log"]
